+++ Infin8.Coapp.Models/Fin_Ledger_Grp.cs
@@ -10,7 +10,7 @@ namespace Infin8.Coapp.Models
         public string? Grp_Name { get; set; }
         public int Fnl_Id { get; set; }
         public bool Grp_Mapped { get; set; }
-        public int Usr_Id { get; set; }
+        public decimal Usr_Id { get; set; }
         public int Grp_SlNo { get; set; }
         public bool Grp_Delete { get; set; }
         public string? BrCode { get; set; }
