+++ Infin8.Coapp.Models/Fin_Ledger_SubGrp.cs
@@ -10,7 +10,7 @@ namespace Infin8.Coapp.Models
         public string? SubGrp_Name { get; set; }
         public int Grp_Id { get; set; }
         public bool SubGrp_Mapped { get; set; }
-        public int Usr_Id { get; set; }
+        public decimal Usr_Id { get; set; }
         public int SubGrp_SlNo { get; set; }
         public bool SubGrp_Delete { get; set; }
         public string? BrCode { get; set; }
